Aws/Lambda/lab01/s3-to-rds-v2/requirements.txt
mysql-connector-python==8.0.33
chalice
boto3
pytest-cov